qmlls: autocompletion for singletons, attached types, enums
Add static helpers enumerationCompletion and enumerationValueCompletion
to provide completion for enumerations types and enumeration values
inside of script identifier expressions.

Extend the LocalSymbolsType enum to be able to query Singleton types
and AttachedTypes
Implement filteredImportedNames to be able to filter imported names by
singleton-ity or attached-ity for example.

Add a semantic scope to QmltypesComponent. They are very similar to
QmlComponents, except that those were created from a .qmltypes file
instead of a .qml file. This allows to query QQmlJSScope's from imported
types, for example when searching for Singletons or Attached types.
Also, this is 'for free' because the qqmldomtypesreader already has a
semantic scope when it constructs the QmltypesComponent. Also add the
semantic scope to the QmlObjects inside the QmltypesComponent.

Fix Export::isSingleton member: Exports obtained from qmltypes file
did not set this member value.

Extend QQmlLSUtils::resolveIdentifierExpressionType() to be able to
resolve enums, singleton types and attached types (for attached
properties).

Implement completion support in qmlls for enumerator names and values,
and Singleton types using the improved
resolveIdentifierExpressionType().

Also autocomplete attached types, and make a difference between
Constructors that represent constructable types and Classes which
represent all kind of types, including attached types. Autocompletion
for properties and co inside the attached types is not supported yet.

Extend the QQmlJSTypeResolver to be able to access all imported types,
needed for searching singletons or attached types for example, and
remove the duplicated DOM methods (duplicated in the sense that they do
the same thing as the QQmlJSTypeResolver).

Task-number: QTBUG-116899
Change-Id: Ibadeb72e2250a6530c42b9d4a7f25105ab6ea31f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>