Fix BusyIndicator invisibility under Fusion style
When an initially invisible BusyIndicator is first made visible under
the Fusion style, it does not in actuality become visible. This is
because the QQuickFusionBusyIndicator::itemChange() function ignores
ItemVisibleHasChanged, so the appropriate actions don't take place. The
fix is a partial copy and paste of
QQuickMaterialBusyIndicator::itemChange's implementation which has
identical logic except for calling the appropriate ::itemChange() super
function.

Task-number: QTBUG-108808
Pick-to: 6.7 6.6 6.5
Change-Id: Id92c62a1eef4fc278ab91097f04db5b41a5d2c8a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>