qmlls: do not autocomplete property definition name
Make QQmlLSUtils::itemsFromTextLocation smarter: it differentiates now
between property definitions and bindings.

Example:

For this property
```
property var somePropertyDef: Item { property var propertyDef: 42; }
```
The ranges of the filelocations are as follows:

somePropertyDef FileLocations range:
|------------------------------------------------------------------|
propertyDef FileLocations range:     |-----------------------------|
binding FileLocations range:                      |----------------|

The previous behavior of QQmlLSUtils::itemsFromTextLocation could
not distinguish between a binding being alone or a binding being in
a property definition, and would return the binding.

The new implementation recognizes when a binding is inside a property
definition, and returns this propertydefinition if the current text
position is before the ':'-token. It even can choose the right property
definition when there are nested ones, like in the example above with
somePropertyDef and propertyDef.

This means that qmlls can now differentiate between being inside a
property definition name and the binding of the property definition.

This allows to suppress completions inside of property definition
names. Completion inside the binding of a property definition are
still handled by the code for normal bindings.

Fixes: QTBUG-117440
Change-Id: If0592ccadd8b1e3e9efbf4952c8044c40854202b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>