QQmlObjectCreator: Don't crash when installing bindings
When creating qqml object, all pending bindings are installed and run
once to determine their dependencies. If they do not have any, they
cannot change and are therefore removed.

A missing check for the ref count of the binding after it was set lead
to a use-after-free. A check was added to determine if the binding has
enough remaining references to be used.

Pick-to: 6.6 6.5 6.2
Fixes: QTBUG-117642
Change-Id: Ida94d46d6f482a5294b19c41578e592b42906634
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>