qmlls: complete handler names in bindings
Previously, qmlls was not able to find property change handlers or
signal handlers for properties or signals defined from QML code. Instead
of trying to fix the DOM code that searches for signals and properties
in the current type, use the qqmljsscope to find properties and signals
of the current type.

Add a test that makes sure that the autcompletion
proposes handler names for QML defined signals and properties.

Change-Id: I07910044a23a2b9df9a6d03fb37be5c06d10b213
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>