StyleGenerator: Add PageIndicator control
Follow the approach of the ComboBox, where the page indicator's delegate
is listed as a separate control in the config file with its own states.
This is because the delegate has not only its own states, but also
context properties (eg: index) that are not accessible to the page
indicator itself.

Change-Id: I7065bc86e094469c50451c693ceeefcfed1338c5
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>