qmlls: implement completion in JS for statement + test snippets
Add the missing source locations to the Dom for the for loop statement.
With this information, the completion knows exactly which part of the
for-loop is being modified and corresponding completion can be provided.
Now that one statement does provide completion for other JS statements,
also test the JS statement snippet completion.

Now that JS statement completion is working in for statements, use it to
test the snippet completion.

Change-Id: I5b0efad09c5617ebe90cbb0418ca04f3ae459de3
Task-number: QTBUG-117445
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>