qmldom: introduce FileLocationRegion enum as key for sourcelocations
Previously in the Dom, SourceLocations and comments were stored in
QMaps and keyed by strings (so-called region names). This commit
replaces the string keys with a newly introduced enum called
FileLocationRegion.

Replace the QMaps in
FileLocations::{regions,preCommentLocations,postCommentLocations} and
RegionComments::{preComments,postComments} to key by FileLocationRegion
enum instead of string, along with all their usages.

The DomItem Dom::Map interface requires the above-mentioned maps to be able to key
by string, so implement two helper methods fileLocationRegionValue() and
fileLocationRegionName() in qqmldom_utils that can translate a string to
FileLocationRegion and back. Those two helpers are used to construct a
Dom::Map in Map::fromFileRegionMap and Map::fromFileRegionListMap.
Add a test for Dom::Map constructed in that way.

Change-Id: I185a4114c9db9f86d9e219079d12d833abfe590e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>