Dom: simplify findAttachedInfo
Remove the options enum from findAttachedInfo to remove all the
std::optional<> from the AttachedInfoLookupResult object. Previously, if
you called findAttachedInfo, you had to pass it a FindOptions enum
value. This FindOptions enum value would decide which fields of the
AttachedInfoLookupResult returned by findAttachedInfo are std::nullopt
or set.

This makes the handling of AttachedInfoLookupResult<> quite annoying as
you have optionals everywhere that just reflected which enum value was
passed to findAttachedInfo.

Now, remove this enum and drop all std::optional and std::nullopt
checking from AttachedInfoLookupResult. Also, split it into
AttachedInfoLookupResultBase and AttachedInfoLookupResult to fix a bug
in AttachedInfoLookupResult::as() where foundTreePath was not copied to
the resulting object. Instead, put all non-template related members into
AttachedInfoLookupResultBase and use the default
AttachedInfoLookupResultBase assignment operator to copy all the members
into the resulting object.

Finally, add a comment in FileLocations::findAttachedInfo to describe
what this AttachedInfoLookupResult can do, and how to print out a
FileLocations::Tree with it (this makes debugging much easier I
believe).

Change-Id: I64a6bd9de11c576ccb34ccfc32ff838f0e903d91
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>