StyleGenerator, Switch: don't include indicator.x in the text padding
The current implementation fails to position the text correctly
in mirrored layouts. The reason is that the contentItem should
not take indicator.x into account since the calculation of
indicator.x uses control.padding explicitly. The contentItem,
OTOH, is already placed according to control.padding implicitly.

So, do the same as we do with e.g RadioButton, and only use
indicator.width and spacing in the calculation.

Change-Id: I8b32f7fbe3757a320ab21a59d0b56b89fb7d46c4
Reviewed-by: Doris Verria <doris.verria@qt.io>