StyleGenerator, Frame: let implicit size use minimum frame size
It's a bit random how big the designer designs the frame in
Figma. So the image we generate from the background can end
up quite big. But regardless of the image size, we should
make sure that the Frame size in controls ends up as tight
as possible around the contentItem.

This patch will therefore introduce minimumWidth/Height in
StyleImage that tells how small the image can be. And this
is the size we should use when calculating the implicit size
of the Frame.

Change-Id: I5e7e195a91b047ac7affc460b1b786dca5c86724
Reviewed-by: Doris Verria <doris.verria@qt.io>