iOS Style: Use separate elements for the track based on the indeterminate state
When the ProgressBar is indeterminate a NumberAnimation runs and
animates the x value of the track. This presents a problem when the
ProgressBar changes from indeterminate to determinate, causing the x
value of the determinate track to start at the last location
the animation stopped.
Since we don't want to use any JS expressions in our templates, fix
this by having a separate track (progress) image for the
indeterminate state and have that be the animation target, instead
of reusing the same element for both states.

Pick-to: 6.5 6.6
Change-Id: I85b338b59ea662c2d6e0afff8fbfeeb381beab19
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>