StyleGenerator, TextArea: let implicit size use background minimum size
The implicit size of a TextArea should not follow the background
image size. Instead it's the size of the contentItem that should
"decide" that implicit size.

This patch will therefore use the background's minimumsize in
the implicit size calulcation. Then it doesn't matter how big
the TextArea was designed in Figma.

Change-Id: Ib85771fedd36e7172cda7862640659f6fd88f5d2
Reviewed-by: Doris Verria <doris.verria@qt.io>