codegenerator: Convert to the original type of registers for unary ops
When generating the code for unary operations, we converted from the
received type to the expected type in readAccumulator. Except, if the
latter was replaced by a different type such as with a generalization to
QVariant for example, then we would not retrieve the original type the
operations should be performed on but keep the replacement type.

Convert the received type to the original type instead of to the
replacement.

Fixes: QTBUG-117789
Change-Id: Ia0109918443b1e1be2bc57b9d46a3a628799806b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d4bab563568a211bf05260fed1b80b4652b20507)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>