Fix regression on singletons where declared != actual type
For singletons, we can't rely on the static meta-object of the
registered type, as the create function used for a singleton might
return an instance of a derived type instead.
This could lead to crashes or method-overload resolution failures later
on.
Fix this by always querying the actual meta-object of the singleton,
instead of relying on the meta-object of the registered class.

Leave a comment about the fact that we still have an issue if you have a
derived instance of the registered type when you additionally extend the
type. However, that is pre-existing (given that it only affects QQmlType
itself).

Fixes: QTBUG-117891
Pick-to: 6.5 6.2
Change-Id: Ie12b4c886ace5cc58a5ded59fa48bc425ae58419
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 8c2ea501fdd07e44ee68e22dcc6d820ae93116f2)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>