StyleGenerator, ProgressBar: add support for padding
In Figma, you can now set a padding on the ContentItem
to control where the track should be placed inside the
groove. This allows the track to have a little bit
of space around it.

In order for ProgressBar.qml to make use of the padding,
this patch will add the ContentItem atom the config file.

Change-Id: Iedd574613bf51008a115903cc93300f9de87ff0f
Reviewed-by: Doris Verria <doris.verria@qt.io>