StyleGenerator, SpinBox: update SpinBox to follow the new Figma layout
The layout for the SpinBox has been simplified in Figma. Implement
the changes needed in the config file and SpinBox.qml to
reflect this.

Especially, the padding in Figma will now only control the
position of the Label. The up and down indicator backgrounds
are always stretched all the way to the top and bottom of
control (and hence, should therefore ignore the top and bottom
padding).

Change-Id: I1de0742eed89eb85543190e2f80646ad6fcea96a
Reviewed-by: Doris Verria <doris.verria@qt.io>