doc: Warn about mixing of implicit/explicit assignment to default props
Assigning objects to default list properties like children for Items can
be done implicitly by defining objects inside the item, or explicitly by
assigning the property directly with an array. Mixing the two solutions
leads to undefined ordering of the objects. Additionally, the behavior
differs between the interpreter and the compiler.

Warn about these issues and signal that both methods must not be mixed
within a single element's definition.

Fixes: QTBUG-102468
Change-Id: Ib9c467358d34105183ef4e20f8b28c687b6db8c5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>