Palette: Remove redundant code
In 4e7a83156d47683f777e92d2790cb617f0bb09d5 we duplicated the
implementation for the bitPosition() function in qpalette.cpp.

The intention was to expose the function in a private header, and
eventually replace the redundant functions in qpalettecolorprovider,
with the ones in QtGui.

Now that qpalette_p.h exist, lets actually use it.

Change-Id: Id7528b7f1c3f5481e0410688ef8e018afb951410
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>