StyleGenerator: GroupBox: Make label have same height as in Figma
We calculate the paddings and positionings based on the properties of
the groupbox background and label in the Figma template.
For the calculations to be accurate and for our control to represent
as closely as possible the Figma template, the GroupBox's label in QML
needs to have the same height as the Label in Figma.

Change-Id: Iaf589ce92f7d0cff4f546afe0c3aef205472bd66
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>