qmltyperegistrar: Fix namespace registration
Whenever the target type is a namespace we have to generate a namespace
registration. This is in line with how we decide what to do with
Q_GADGET and Q_OBJECT types. Furthermore, even if both types are not
namespaces, but we have a QML_FOREIGN_NAMESPACE marker, we still need to
generate a namespace registration. The normal type registration would
just ignore the foreign. To this end, add another type info that tells
us whether the foreign type is supposed to be a namespace. Finally, also
add the override marker to QML_FOREIGN_NAMESPACE. We can derive from
types that declare it after all.

Pick-to: 6.5
Fixes: QTBUG-117703
Change-Id: I6ab25989f7738ab4567f0748192449704031bcc6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 1be6d083a22a96f6b21a2eb5d7fc87a67ca3b080)
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>