StyleGenerator: fixup config according to recent changes in Figma
The template structure in Figma has changed. Instead of
having a ContentItem with a layout, the ContentItem has
been removed, and the layout has been placed on the root
item instead. At the same time, the names of the
primitives has also been simplified to make all the
templates look the same (which also makes them easier to
document and understand).

This patch will update the config file to reflect those
changes.

Change-Id: I7a46dfbc48ccce33c6f9963a24342e34653a86af
Reviewed-by: Doris Verria <doris.verria@qt.io>