Fix translucent NativeRendering text on transparent window
We cannot pass the color alpha to the source color
scale factor, since when we blend we will end up with a
resulting output color that has alpha. The color's alpha is
premultiplied into the subpixel alpha values we fetch in the
fragment shader, together with the state's opacity.

Note: While not strictly necessary, this also changes the
text shader to match the standard premultiplication in other
shaders, since it has caused some confusion.

[ChangeLog][Text] Fixed an issue where NativeRendering text
with a translucent color combined with a transparent window
would cause unexpected translucency effects on opaque parts
of the scene.

Pick-to: 6.5 6.6
Fixes: QTBUG-118052
Change-Id: I1f06a503c48355863784fc18a849fb539b9384ef
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>