QQuickCheckBox: make nextCheckState property non-private
QQuickCheckBox has private properties but its private class
QQuickCheckBoxPrivate is defined in a .cpp instead of a header.
This causes problems with qmltc because it needs to include the
header of the private class when it sees private properties.

Replace the private property nextCheckState with a non-private one, as
nextCheckState does not need to be private (QJSValue is public API).
Amends 043e3d24f9306c18af97156ec613bb64a9dc0d91 who exposed this
property to QML.

Add a file with a CheckBox in the qmltc tests to test if qmltc can
properly compile a CheckBox now.

Note: call the getter of nextCheckState "getNextCheckState" to not
shadow the "nextCheckState" virtual method from the base class.

Fixes: QTBUG-118091
Pick-to: 6.5
Change-Id: I771aeb8a4f388d7068ec78eba968746c8e184f31
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 8ba3d83f28a5d4c84a13df001d59cbcb78ba0f2f)