qmlTypeId: Import unregistered modules
When using the imperative registration API, one could first register all
types and then query the engine for their id via qmlTypeId. With the
declarative API, there is no such "safe point" (short of creating a QML
engine, and importing the module there).

We can't really avoid the creation of an engine (something has to import
the module before it becomes available), but we can hide it in the
implementation of qmlTypeId, so that the function "just works™".
This uses the same logic as QQmlComponent::loadFromModule behind the
scenes, but stops short of creating the object, as we only need the
QQmlType for its id.

Pick-to: 6.7 6.5
Fixes: QTBUG-118165
Change-Id: Iec456140c32c036ae91db62be5a598ffc8b34607
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 336efa68a3624d2c8e1d28ec445bcf261318acc0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>