StyleGenerator: Add entry for default controls
Some controls are not part of the Figma templates but we need to add
their QML implementations to the Figma style so that we can use them.

They need to be added to the qmldir file together with the rest of
the controls when generating the style. To do that, add a new
entry "default controls" to the config file. We want to keep them
separate from the other controls found in the "controls" array
because we want to enforce required keys such as "component set",
"atoms", etc for those controls, which don't make sense for the
default ones not found in the Figma templates.

Change-Id: Ib6b94a191a355d2a5d676464199e77901006fe9f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>