tst_qmlls_modules: skip buildDir test
Add a QSKIP() in a tst_qmlls_modules::buildDir that started
failing because of eade144a6d95a715decfd7ab7b1fd7c164041c9. That was
never caught by CI because CI was not running this test prior to this
commit.

Note: it was not possible to QEXPECT_FAIL the test because it fails an
assertion inside a helper method shared with other unittests (and that
would QEXPECT_FAIL other unrelated tests that actually pass).

Created QTBUG-118424 to track the issue in buildDir.

Task-number: QTBUG-118391
Change-Id: I0fde124b979bed22a893e195e0d354fd9b6e8b0d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>