qmldom: Dom representation for dot separated identifiers
We were keeping attached or grouped property names in Dom::Binding
as a single string in form of "id.property" or "AttachedObj.property"
which was not suitable to use in helper functions in qmlls utils.
Add a new field in Dom::Binding to represent bindingIdentifier in the
form of ScriptBinaryExpression<ScriptIdentifier, ScriptIdentifier>
to for grouped/attached properties or ScriptIdentifier for the normal
binding names which doesn't contain a dot..

resolveExpressionType can find the same scope more than one due to the
new fiels, do not add location results if they were already added.

Task-number: QTBUG-117380
Change-Id: Iccaf74095356fa181d1936a48e92316069e50051
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>