QtQml: Use array data for QmlListWrapper's "owned" mode
We don't want to get surprised by QObjects being deleted while we're
still holding on to them. Also, we want to mark them when gc'ing and we
want to use the write barrier when replacing them. We already have an
array data since QmlListWrapper is an object. Let's use it.

Pick-to: 6.6 6.5
Task-number: QTBUG-117793
Change-Id: I921514bc2de994884c41bd6af19a7c2ac746cd92
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>