tst_qmlls_modules: skip rangeFormatting test
It seems that newlines are getting mixed up on windows during
rangeformatting, at least in the tst_qmlls_modules::rangeFormatting
tests. QEXPECT_FAIL those tests on windows, and fix the typo in the
rowname.

That was never caught by CI because CI was not running this test before.

Created QTBUG-118425 to track this issue.

Task-number: QTBUG-118391
Change-Id: I5264fa63d3a1102b4edd40aec82b7e1eec784328
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>