MessageDialog: make QQuickAbstractDialog::result return button pressed
The widgets QMessageBox returns either the StandardButton value of the
last button pressed, or the index, in the case of custom buttons.
This differs from how we did things in quick, where we'd always return
either 0 or 1 (Accepted or Rejected).

The documentation even states the following:

MessageDialog sets the result to the value of the clicked standard
button instead of using the standard result codes

Which was incorrect, since the result property functioned the same,
regardless of dialog type.

The goal of this patch is to make the result property for MessageDialog
behave the same as QMessageBox, without changing the behavior for other
dialog types. In order to accomplish this, the type has to be changed
from StandardCode to int, since the result value can either be a
StandardCode value, or a StandardButton value.

The widget QMessageBox also differs in that it will not emit rejected
when the user presses a button with a role that isn't NoRole or
RejectRole. I've decided to make the MessageDialog behave the same way.

[ChangeLog][QtQuick][Dialogs][Important Behavior Changes] The
MessageDialog will no longer emit rejected when the user presses a
button that doesn't have the role NoRole or RejectedRole.

Task-number: QTBUG-118445
Task-number: QTBUG-118689
Task-number: QTBUG-118212
Change-Id: I7084a889210027f5c057f5c22eee2e08867ce741
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>