tst_qmlls_modules: make the buildDir test run in the CI
Amends eade144a6d95a715decfd7ab7b1fd7c164041c9 that forgot to adapt the
test to the new completion.
Change the completion category to Constructor for QML Objects
completions and simulate a change of the QML file to make the Dom
recreate the DomItem with the updated build directory information.

Fixes: QTBUG-118424
Change-Id: I4492c1e2a143048fd25af80a02611a3018f13318
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>