StyleGenerator: Generate qml config file instead of json
By dumping the configuration directly into a qml file we skip an
unnecessary (second) JSON parsing step and converting from JS to
C++. Also whenever the theme changed, we were reparsing the config
file since we have a different configuration for each theme.
So this should speed things up quite a bit.
Make the generated QML a singleton and include it in the qmldir as
such.

Change-Id: I7473d9f6f9977eae5b9d19534c99e8bd9aa6e4b9
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>