qmllint: do not construct well known valuetypes from string (part 1)
Add a warning when well-known QML_STRUCTURED_VALUE's type are
constructed from strings in bindings, with a fix-it to
automatically generate the QML_STRUCTURED_VALUE way of constructing
value types.
Do not provide a fix-it in the case that a value type is constructed
from an invalid string, like "50x70" instead of
"50,70" for a QPointF, for example.

Previously, qmlcachegen was compiling some of the string assignments for
structured types like points, and some not, like vector2d. In the latter
case, it printed the generic "Cannot assign literal of type ..." from
the qmlIncompatible category.

This commit only handles point, size and rect. A later commit will take
care of the quick/gui types matrix4x4, vectorNd and quaternion.

Now that the methods are all accessible from one point and consistently
reused, implement qqmljsvaluetypefromstringcheck that checks if a given
type should be constructed from a stringliteral, and generates a
suggestion using the new way of constructing structured value types if
it can. This is implemented into a separate header because it might be
needed for multiple checks.
Currently, it is only needed for the stringliteral "binding" check,
but it will also be needed for stringliteral "assignment" checks,
where you warn about `myPoint ="1,2"`, for example (see QTBUG-118547).

To avoid code-duplication in the quicklintplugin, extract the code also
needed for the qquickliteralbindingcheck into a separate base class
LiteralBindingCheckBase, such that qquickliteralbindingcheck, in a later
commit, may just extend LiteralBindingCheckBase using its
isValueTypeFromStringConstruction override.

Task-number: QTBUG-118323
Fixes: QTBUG-118546
Change-Id: I8572798d94593b1044cdb49116c8468f998dbbd7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>