qmllint: warn when constructing quick valuetypes from string
This commit takes care of care of the quick/gui types matrix4x4,
vectorNd and quaternion.

Add a warning when well-known QML_STRUCTURED_VALUE's type are
constructed from strings in bindings, with a fix-it to
automatically generate the QML_STRUCTURED_VALUE way of constructing
value types.
Do not provide a fix-it in the case that a value type is constructed
from an invalid string, like "50x70" instead of
"50,70" for a QPointF, for example.

Use the LiteralBindingCheckBase to avoid code duplication in
qquickliteralbindingcheck, and provide the check for the quick types via
the overridden method.

Add a test for tst_qmllint.

Task-number: QTBUG-118323
Fixes: QTBUG-118546
Change-Id: Ibfda260f011f7683378771ca280ad70b0bd0fcfa
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>