qmlls: find usages of group properties
Add support for finding usages of the grouped properties. A grouped
property might be in the form of a.b or a{}. In the latter case, it is
represented as QmlObject in the dom. Add QmlObject handling to
findUsagesOfNonJSIdentifiers method, also handle that case while
resolving expression types.

QQmlJSMetaPropertyBinding has provides an api to get group property
type. Add an helper to query the scope from propertyBindings.

The following commits will handle the attached property, enum or
singleton usages. Another separate work is still needed for generalized
group properties.

Task-number: QTBUG-117380
Change-Id: I0cae7ba1e2bc9191b1e69ee8729aa509ba7fd5b9
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>