Fix compile time qsTranslate with empty context
An empty context is to be passed as-is. We shall not replace it with the
file context context. Since the TranslationData struct has a field for
the context, we need to invent a "no context" value we use for the
methods that don't allow you to set a context (e.g. qsTr, qsTrId). We
cannot use 0 because that is the empty string which is a valid context
now.

Amends commit 9cfc19faf5d1ce2b9626914ab4528998b072385d.

Pick-to: 6.5
Fixes: QTBUG-118469
Change-Id: I160c512f42aba4a8ae2fc8860cdf4e50c53d9d3e
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 5fc4c09a958c5f78aa9f69cee2c45eef9a5166b8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>