qmllint: Add warnings to importFromHost
When importing, QQmlJSImportVisitor creates a url from the filename.
In cases when URL schema is empty and url is not Qrc
it tries to import from Host.

This commit adds a warning and early exit if file or directory
doesn't exist.
It also adds a warning if the imported object exists,
but it's neither a file nor a directory

Task-number: QTBUG-118074
Change-Id: I59cea90d611a68f2e662040a5ee4ba0ae3f0fe5b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>