Fix crash when calling QQmlIncubator::abort()
QQuickContainerPrivate::itemDestroyed did not properly
clean up its internal data structures to have no references to the item
that got destroyed. This was mainly because the call to

  q->isContent(item)

in QQuickContainerPrivate::removeItem() would return false, so it would
not continue doing the rest of the required cleanup that removeItem
does, and it was therefore leaving dangling pointers to the destroyed
item in QQuickContainers data structures.

Pick-to: 6.5 6.6
Fixes: QTBUG-118397
Change-Id: I7f23b4f98df33d69b42455d20016adb7e132b510
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>