QmlDom: Do not copy an atomic into a lambda
This did not work anyway. There was no synchronization going on because
every time the lambda was copied (which was often), the atomic was also
copied. Let's be honest about this and use a regular integer. If you
call the same copy of the callback toLoad times it will decrement to 0,
as before. Otherwise it does something strange.

Change-Id: Iec8fe8f337e9e9ab58b2f8cfc3d824695b3281eb
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>