dom: allow script elements to have file location regions
Extend the ScriptElementBase class to contain subregions. Like this,
script elements can store source locations in the FileLocations::Tree,
in the same way as the qml dom elements do.
This is required to provide completion in JS statements, as those need
to know if the cursor is before or after a certain token or keyword to
provide completion at the right places.

Task-number: QTBUG-117445
Change-Id: Ifc28b35da9371c780e1d3d6ea4cd3974bcca30fd
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>