StyleGenerator: add support for theme icons
Generate the theme icons from the figma design.

The icons are generated as theme icons, which
means that the application will need to add the
following line to main.cpp in order to use them:

QIcon::setThemeName("NameOfStyle");

The icons can then be used in e.g Buttons by using
the name of the icon:

Button {
    icon.name: "navigation_settings"
}

The name used is the same name as the name of the
variant in Figma, but with "_" between each category.

Change-Id: I58110f0d74a95ae838271d28f56a2a912b82dac9
Reviewed-by: Doris Verria <doris.verria@qt.io>